fixes to view rest framework static files
The reason for this change was when running app via gunicorn files
from the static folder such as css and js would not be served up.
The change enables django rest framework to serve its static files
in development mode. To serve static files in production mode it
is advised to use nginx to handle static files and gunicorn to
handle the application specific files.