bug fix to timestamps when close/reopen/delete/amend defect
The reason to this bug is due to the default=timezone.now()
within the method args. What actually happens is this
argument is not re-evaluated each time during runtime but
only once during compile time instead (the time main python
executable is first run). What ends up happening is that
the same timestamp is used every time the method is called.