some small hack to get import working again
The tradeoff with this change is that an explicit import event does not get saved to the database. Instead the default 'defect opened' event will get saved by the signal handler on post create defect.

For future changes, perhaps a defect imported event can get saved after the the defect opened event with an empty payload. However this will require changed behaviour of the domain model and updates to previous 'defect imported' events for consistency.